+++ app/store/configureStore.dev.js
@@ -8,7 +8,7 @@ import Store from 'electron-store';
 
 import throttle from 'lodash/throttle';
 import rootReducer from '../reducers';
-import rootSaga from '../sagas';
+import rootSaga from '../core/sagas';
 import { saveState, loadState } from './localStorage';
 
 const history = createHashHistory();
